% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearModelSummary.R
\name{linearmodel}
\alias{linearmodel}
\title{The linear model function will fit a linear model using plot_id and weight.}
\usage{
linearmodel(data)
}
\arguments{
\item{data}{Data that is being used to run the function /cloud/project/hanhart.pkg/data/surveys.csv}
}
\value{
The output, in this case the linear model summary, of which the function returns /cloud/project/hanhart.pkg/data/lm.txt
}
\description{
Produce a summary of a linear model with plot_id and weight, returning insight of residuals, error rates, and estimates.
}
