% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointPlot.R
\name{pointplot}
\alias{pointplot}
\title{The point plot function will create a plot using geom_point}
\usage{
pointplot(x = weight, y = species_id, surveys)
}
\arguments{
\item{data}{Data that is being used to run the function /cloud/project/hanhart.pkg/data/surveys.csv}
}
\value{
The point plot created by the function /cloud/project/hanhart.pkg/data/pointplot.pdf
}
\description{
Returns a point plot of weight by species_id, grouped by sex.
}
